fees to bt pkg new input rename
