+++ bt/input/Input.go
@@ -31,8 +31,8 @@ type Input struct {
 	SequenceNumber     uint32
 }
 
-// New creates a new empty Input object with a finalised sequence number.
-func New() *Input {
+// NewInput creates a new empty Input object with a finalised sequence number.
+func NewInput() *Input {
 	b := make([]byte, 0)
 	s := script.NewFromBytes(b)
 
@@ -42,8 +42,8 @@ func New() *Input {
 	}
 }
 
-// NewFromBytes returns a transaction input from the bytes provided.
-func NewFromBytes(bytes []byte) (*Input, int, error) {
+// NewInputFromBytes returns a transaction input from the bytes provided.
+func NewInputFromBytes(bytes []byte) (*Input, int, error) {
 	if len(bytes) < 36 {
 		return nil, 0, fmt.Errorf("input length too short < 36")
 	}
@@ -71,8 +71,8 @@ func NewFromBytes(bytes []byte) (*Input, int, error) {
 	return &i, totalLength, nil
 }
 
-// NewFromUTXO returns a transaction input from the UTXO fields provided.
-func NewFromUTXO(prevTxID string, prevTxIndex uint32, prevTxSats uint64, prevTxScript string, nSeq uint32) (*Input, error) {
+// NewInputFromUTXO returns a transaction input from the UTXO fields provided.
+func NewInputFromUTXO(prevTxID string, prevTxIndex uint32, prevTxSats uint64, prevTxScript string, nSeq uint32) (*Input, error) {
 	pts, err := script.NewFromHexString(prevTxScript)
 	if err != nil {
 		return nil, err
